Break out admin into separate components
