+++ sass/components/Admin.module.scss
@@ -19,165 +19,12 @@
   margin-bottom: 1rem;
 }
 
-.selectImageLabel {
-  position: relative;
-  left: 1.4rem;
-}
-
-.imagePartSelect {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 2rem;
-
-  select {
-    padding-top: 0.3rem;
-    padding-bottom: 0.3rem;
-    padding-left: 0.3rem;
-    width: fit-content;
-  }
-}
-
-.editorIsDraftCheckbox {
-  margin-bottom: 1rem;
-  margin-top: 1rem;
-
-  label {
-    margin-right: 0.5rem;
-  }
-}
-
-.editorMainImage {
-  max-width: 28rem;
-  margin: auto;
-}
-
-.editorMainImage {
-  display: flex;
-}
-
-.editorTitleContainer {
-  display: flex;
-  justify-content: space-between;
-  #editorSaveButton {
-    font-size: 0.8rem;
-  }
-  #previewButton {
-    font-size: 0.8rem;
-  }
-}
-
-.editorPostShortIdTagGroup {
-  span {
-    font-size: 0.88rem;
-
-  }
-  margin-bottom: 1.8rem;
-  margin-top: 1rem;
-}
-
 .expandPostArrow {
   cursor: pointer;
   position: relative;
   font-size: 1.5rem;
   right: -1rem;
   top: 0.24rem;
-}
-
-.mdPartEditor {
-  margin-bottom: 1.3rem;
-}
-
-.tweetSinglePartEditor {
-  label {
-    display: block;
-  }
-
-  input {
-    &::placeholder {
-      color: rgb(61, 194, 255);
-    }
-    color: rgb(117, 204, 255);
-    background-color: rgb(37, 37, 37);
-    margin-right: 0.6rem;
-    width: 300px;
-    padding-left: 0.4rem;
-    padding-bottom: 0.4rem;
-    padding-top: 0.4rem;
-  }
-}
-
-$dot-size: 0.9rem;
-.redDotDeleteMdPart {
-  height: $dot-size;
-  width: $dot-size;
-  background-color: rgb(248, 66, 66);
-  border-radius: 50%;
-  display: inline-block;
-  cursor: pointer;
-
-  position: relative;
-  top: 1.9rem;
-  left: 0.5rem;
-}
-
-.redDotDeleteImgPart{
-  height: $dot-size;
-  width: $dot-size;
-  background-color: rgb(248, 66, 66);
-  border-radius: 50%;
-  display: inline-block;
-  cursor: pointer;
-
-  position: relative;
-  top: 1.2rem;
-  left: 0.1rem;
-}
-
-.redDotDeleteTweetPart{
-  height: $dot-size;
-  width: $dot-size;
-  background-color: rgb(248, 66, 66);
-  border-radius: 50%;
-  display: inline-block;
-  cursor: pointer;
-
-  position: relative;
-  top: 0.1rem;
-  left: -0.4rem;
-}
-
-.textAreaEditor {
-  margin-bottom: -0.5rem;
-  font-size: 0.99rem;
-  padding: 1rem;
-  padding-top: 2rem;
-  width: 100%;
-  height: 20vw;
-  border: solid rgb(202, 202, 202) 0.03rem;
-  border-radius: 0.33rem;
-  border-top-width: 1.9rem;
-  border-top-color: rgb(236, 236, 236);
-  
-}
-
-.categoryInput {
-  font-size: 1.23rem;
-  padding-left: 1rem;
-  padding-bottom: 0.5rem;
-  padding-top: 0.5rem;
-  width: 100%;
-}
-
-.titleInput {
-  font-size: 1.23rem;
-  padding-left: 1rem;
-  padding-bottom: 0.5rem;
-  padding-top: 0.5rem;
-  width: 100%;
-}
-
-.titleEditing {
-  font-size: 2rem;
 }
 
 .postCardListScrollableContainer {
@@ -215,10 +62,6 @@ $dot-size: 0.9rem;
   justify-content: space-evenly;
   margin-left: 0.3rem;
   margin-right: 0.3rem;
-
-  // button {
-  //   width: 100%
-  // }
 }
 
 .postCardImagesButton {
@@ -253,10 +96,6 @@ $dot-size: 0.9rem;
 
 .postCardDeleteButton {
   width: fit-content;
-}
-
-.postImagesContainer {
-  display: flex;
 }
 
 $dot-size: 0.4rem;
