+++ sass/components/DeploymentControls.module.scss
@@ -1,0 +1,3 @@
+.tagGroup {
+  margin-top: 2rem;
+}
