+++ sass/components/ImageUploadEditor.module.scss
@@ -1,0 +1,3 @@
+.postImagesContainer {
+  display: flex;
+}
